Improvement file organization of the whole project.
