+++ README.md
@@ -3,3 +3,6 @@
 ## Installation
 * setup `venv` 3.7+
 * run `python -m pip install -r requirements.txt` for `venv`
+
+## Run
+* run `python -m scheduler`
