+++ scheduler/Utils.py
@@ -1,0 +1,17 @@
+import os
+
+def prepare_results_directory():
+    """
+    Prepare the results directory for storing output files and clean up old results.
+    """
+    results_dir = "results"
+    os.makedirs("results", exist_ok=True)
+    
+     # Iterate over all the files in the directory and remove them
+    for filename in os.listdir(results_dir):
+        file_path = os.path.join(results_dir, filename)
+        try:
+            if os.path.isfile(file_path):
+                os.remove(file_path)
+        except Exception as e:
+            print(f'Failed to delete {file_path}. Reason: {e}')
