+++ scheduler/methods/Michigan.py
@@ -5,7 +5,7 @@ import numpy as np
 import pandas as pd
 from sklearn.utils import shuffle
 
-import Common
+import scheduler.Common as Common
 
 columns = ''
