+++ scheduler/methods/Pitt_direct.py
@@ -14,7 +14,7 @@ import random
 import csv
 from collections import Counter
 
-import Common
+import scheduler.Common as Common
 
 ITERATIONS_NUMBER = 100
 POPULATION_SIZE = 10
